Enum spirv_types::image_params::ImageFormat

#[repr(u32)]
pub enum ImageFormat {

    Unknown,
    Rgba32f,
    Rgba16f,
    R32f,
    Rgba8,
    Rgba8Snorm,
    Rg32f,
    Rg16f,
    R11fG11fB10f,
    R16f,
    Rgba16,
    Rgb10A2,
    Rg16,
    Rg8,
    R16,
    R8,
    Rgba16Snorm,
    Rg16Snorm,
    Rg8Snorm,
    R16Snorm,
    R8Snorm,
    Rgba32i,
    Rgba16i,
    Rgba8i,
    R32i,
    Rg32i,
    Rg16i,
    Rg8i,
    R16i,
    R8i,
    Rgba32ui,
    Rgba16ui,
    Rgba8ui,
    R32ui,
    Rgb10A2ui,
    Rg32ui,
    Rg16ui,
    Rg8ui,
    R16ui,
    R8ui,
    R64ui,
    R64i,

}

The underlying internal representation of the image.

Variants

Unknown

Representation not known at compile time.

Rgba32f

RGBA channels, 32 bit floating point per channel.

Rgba16f

RGBA channels, 16 bit floating point per channel.

R32f

Single red channel, 32 bit floating point.

Rgba8

RGBA channels, 8 bit unsigned normalized integer per channel.

Rgba8Snorm

RGBA channels, 8 bit signed normalized integer per channel.

Rg32f

Red+Green channels, 32 bit floating point per channel.

Rg16f

Red+Green channels, 16 bit floating point per channel.

R11fG11fB10f

32 bits containing two 11 bit floating point numbers for the Red and Green channels, and a 10 bit floating point number for the Blue channel.

R16f

Single red channel, 16 bit floating point.

Rgba16

RGBA channels, 16 bit unsigned normalized integer per channel.

Rgb10A2

32 bits containing three 10 bit unsigned normalized integers for the Red, Green, and Blue channels, and a 2 bit unsigned normalized integer for the Alpha channel.

Rg16

Red+Green channels, 16 bit unsigned normalized integer per channel.

Rg8

Red+Green channels, 8 bit unsigned normalized integer per channel.

R16

Single red channel, 16 bit unsigned normalized integer.

R8

Single red channel, 8 bit unsigned normalized integer.

Rgba16Snorm

RGBA channels, 16 bit signed normalized integer per channel.

Rg16Snorm

Red+Green channels, 16 bit signed normalized integer per channel.

Rg8Snorm

Red+Green channels, 8 bit signed normalized integer per channel.

R16Snorm

Single red channel, 16 bit signed normalized integer.

R8Snorm

Single red channel, 8 bit signed normalized integer.

Rgba32i

RGBA channels, 32 bit signed integer per channel (not normalized).

Rgba16i

RGBA channels, 16 bit signed integer per channel (not normalized).

Rgba8i

RGBA channels, 8 bit signed integer per channel (not normalized).

R32i

Single red channel, 32 bit signed integer (not normalized).

Rg32i

Red+Green channels, 32 bit signed integer per channel (not normalized).

Rg16i

Red+Green channels, 16 bit signed integer per channel (not normalized).

Rg8i

Red+Green channels, 8 bit signed integer per channel (not normalized).

R16i

Single red channel, 16 bit signed integer (not normalized).

R8i

Single red channel, 8 bit signed integer (not normalized).

Rgba32ui

RGBA channels, 32 bit unsigned integer per channel (not normalized).

Rgba16ui

RGBA channels, 16 bit unsigned integer per channel (not normalized).

Rgba8ui

RGBA channels, 8 bit unsigned integer per channel (not normalized).

R32ui

Single red channel, 32 bit unsigned integer (not normalized).

Rgb10A2ui

32 bits containing three 10 bit unsigned integers for the Red, Green, and Blue channels, and a 2 bit unsigned integer for the Alpha channel.

Rg32ui

Red+Green channels, 32 bit unsigned integer per channel (not normalized).

Rg16ui

Red+Green channels, 16 bit unsigned integer per channel (not normalized).

Rg8ui

Red+Green channels, 8 bit unsigned integer per channel (not normalized).

R16ui

Single red channel, 16 bit unsigned integer (not normalized).

R8ui

Single red channel, 8 bit unsigned integer (not normalized).

R64ui

Single red channel, 64 bit unsigned integer (not normalized).

R64i

Single red channel, 64 bit signed integer (not normalized).